Fixed some warnings.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@9388 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/LibraryBuilder.cpp
@@ -983,7 +983,7 @@ bool CLibraryBuilder::DetectVirtualLAME(HANDLE hFile, QWORD& nOffset, QWORD& nLe
     else
         nSampleRate = samplerate_table[ nId ][ nSampleRateIndex ];
 	int nFrameSize = ( ( nId + 1 ) * 72000 * nBitrate ) / nSampleRate;
-	if ( nFrameSize > nLength )
+	if ( nFrameSize > (int)nLength )
 		return false;
  
 	int nVbrHeaderOffset = GetVbrHeaderOffset( nId, nMode );
